2026-05-4

+++ src/main/java/com/naver/sth/board/notice/NoticeController.java
@@ -12,10 +12,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.naver.sth.board.BoardDTO;
+import com.naver.sth.file.FileDTO;
+import com.naver.sth.member.MemberDTO;
 import com.naver.sth.pager.Pager;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Controller
@@ -30,6 +34,21 @@ public class NoticeController {
 		
 		return "notice";
 	}
+	@GetMapping("down")
+	public String fileDown(NoticeFileDTO noticeFileDTO,Model model)throws Exception {
+		
+		FileDTO fileDTO=noticeService.fileDetail(noticeFileDTO);
+		
+		model.addAttribute("fileDTO", fileDTO);
+		
+		return "fileDownView";
+		
+		
+	}
+	
+	
+	
+	
 	
 	@GetMapping("list")
 	public String list(Pager pager,Model model)throws Exception {
@@ -87,18 +106,36 @@ public class NoticeController {
 		
 	}
 	@PostMapping("update")
-	public String update(NoticeDTO noticeDTO,@RequestParam("attach") MultipartFile [] attach)throws Exception{
+	public ModelAndView  update(NoticeDTO noticeDTO,@RequestParam("attach") MultipartFile [] attach,Model model)throws Exception{
 		
 		int result = noticeService.update(noticeDTO, attach);
-	
-		return "redirect:./list";
+		
+		ModelAndView mv = new ModelAndView();
+		mv.setViewName("redirect:./list");
+		mv.addObject("dto", noticeDTO);
+		
+		return mv;
 	}
 	
 	@PostMapping("delete")
-	public String delete(NoticeDTO noticeDTO) throws Exception {
-		int result = noticeService.delete(noticeDTO);
+	public String delete(NoticeDTO noticeDTO, HttpSession session,Model model) throws Exception {
+		MemberDTO memberDTO=(MemberDTO)session.getAttribute("member");
+		BoardDTO boardDTO=noticeService.detail(noticeDTO);
+		if(boardDTO.getBoardWriter().equals(memberDTO.getUsername())) {
+			int result = noticeService.delete(noticeDTO);
+			return "redirect:./list";
+			
+		}else{
+			model.addAttribute("result", "작성자가 아니다");
+			model.addAttribute("url", "./list");
+			
+			return "commons/result";
+			
+		}
 		
-		return "redirect:./list";
+		
+		
+		
 		
 	}
